database save and load - the load is having issues, need to fix

+++ YManage/src/org/ymca/tvc/ymanage/shared/AttendanceRecord.java
@@ -36,6 +36,6 @@ public class AttendanceRecord implements IsSerializable, Comparable<AttendanceRe
 
 	@Override
 	public int compareTo(AttendanceRecord other) {
-		return Integer.compare(this.meetingId.getNum(), other.getMeetingId().getNum());
+		return Long.compare(this.meetingId.getNum(), other.getMeetingId().getNum());
 	}
 }
